requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
backtrader==1.9.76.123
backtrader-plotting==1.1.0
black==20.8b1
bleach==3.3.0
bokeh==2.0.2
brotlipy==0.7.0
certifi==2020.6.20
click==7.1.2
conda==4.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
distro==1.5.0
entrypoints==0.3
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.3.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
markdown2==2.4.0
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
nltk==3.5
notebook==6.2.0
numpy==1.20.1
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
powerline-status==2.7
prometheus-client==0.9.0
prompt-toolkit==3.0.17
ptyprocess==0.7.0
pycosat==0.6.3
Pygments==2.8.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyreadstat==1.1.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
regex==2020.11.13
ruamel-yaml==0.15.87
scipy==1.6.1
seaborn==0.11.1
Send2Trash==1.5.0
tabula-py==2.2.0
terminado==0.9.3
testpath==0.4.4
toml==0.10.2
torch==1.8.0
torchaudio==0.8.0
torchvision==0.9.0
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==2.0.1
xlwt==1.3.0
